Split the multitude of inner functions within Lorre into modules
 * separate processing of different entities
 * rename the TezosNodeOperations to TezosIndexedDataOperations to
   mark the difference with almost same-named TezosNodeOperator
